+++ pystac_client/client.py
@@ -1,4 +1,5 @@
 from copy import deepcopy
+import os
 from typing import Callable, Optional
 from urllib.request import Request
 
@@ -75,13 +76,13 @@ class Client(pystac.Catalog, STACAPIObjectMixin):
         return '<Catalog id={}>'.format(self.id)
 
     @classmethod
-    def open(cls, url, headers=None):
+    def open(cls, url=None, headers=None):
         """Alias for PySTAC's STAC Object `from_file` method
 
         Parameters
         ----------
-        url : str
-            The URL of a STAC Catalog
+        url : str, optional
+            The URL of a STAC Catalog. If not specified, this will use the `STAC_URL` environment variable.
 
         Returns
         -------
@@ -89,14 +90,23 @@ class Client(pystac.Catalog, STACAPIObjectMixin):
         """
         import pystac_client.stac_io
 
+        if url is None:
+            url = os.environ.get("STAC_URL")
+
+        if url is None:
+            raise TypeError(
+                "'url' must be specified or the 'STAC_URL' environment variable must be set.")
+
         def read_text_method(url):
             request = Request(url, headers=headers or {})
             return pystac_client.stac_io.read_text_method(request)
 
         old_read_text_method = STAC_IO.read_text_method
         STAC_IO.read_text_method = read_text_method
-        catalog = cls.from_file(url)
-        STAC_IO.read_text_method = old_read_text_method
+        try:
+            catalog = cls.from_file(url)
+        finally:
+            STAC_IO.read_text_method = old_read_text_method
         catalog.headers = headers
         return catalog